computeInitialQuantity moved
