+++ src/pages/Facility/services/pharmacy/MedicationBillForm.tsx
@@ -125,6 +125,7 @@ import {
 } from "@/types/emr/medicationDispense/medicationDispense";
 import medicationDispenseApi from "@/types/emr/medicationDispense/medicationDispenseApi";
 import {
+  computeInitialQuantity,
   DoseRange,
   MedicationRequestDispenseStatus,
   MedicationRequestDosageInstruction,
@@ -137,11 +138,8 @@ import { InventoryRead } from "@/types/inventory/product/inventory";
 import inventoryApi from "@/types/inventory/product/inventoryApi";
 import { ProductKnowledgeBase } from "@/types/inventory/productKnowledge/productKnowledge";
 import {
-  add,
-  divide,
   isGreaterThan,
   isZero,
-  multiply,
   round,
   roundWhole,
   zodDecimal,
@@ -929,88 +927,6 @@ export default function MedicationBillForm({
       });
     });
   }, [medications.length, append, form, prescriptionId]);
-
-  function computeInitialQuantity(medication: MedicationRequestRead) {
-    const instruction = medication.dosage_instruction[0];
-    if (!instruction) {
-      return "0";
-    }
-
-    const doseValue = instruction.dose_and_rate?.dose_quantity?.value;
-    if (!doseValue) {
-      return "0";
-    }
-
-    const unitCode = instruction.dose_and_rate?.dose_quantity?.unit?.code;
-    const nonVolumetric = ["{tbl}", "{count}"];
-
-    if (unitCode && !nonVolumetric.includes(unitCode)) {
-      return "";
-    }
-
-    if (instruction.as_needed_boolean) {
-      return round(doseValue);
-    }
-
-    const repeat = instruction.timing?.repeat;
-    if (!repeat?.bounds_duration || !repeat.period_unit) {
-      return doseValue;
-    }
-
-    const convertToHours = (value: string, unit: string) => {
-      switch (unit) {
-        case "h":
-          return new Decimal(value);
-        case "d":
-          return multiply(value, 24);
-        case "wk":
-          return multiply(value, 24 * 7);
-        case "mo":
-          return multiply(value, 24 * 30);
-        case "a":
-          return multiply(value, 24 * 365);
-        default:
-          return 0;
-      }
-    };
-
-    const {
-      frequency = 1,
-      period = "1",
-      period_unit,
-      bounds_duration,
-    } = repeat;
-
-    const totalDurationInHours = convertToHours(
-      bounds_duration.value,
-      bounds_duration.unit,
-    );
-    const periodInHours = convertToHours(period, period_unit);
-
-    if (periodInHours === 0) {
-      return doseValue;
-    }
-
-    const doseIntervalInHours = divide(periodInHours, frequency);
-
-    if (isZero(doseIntervalInHours)) {
-      return doseValue;
-    }
-
-    const numberOfDoses = divide(
-      totalDurationInHours,
-      doseIntervalInHours,
-    ).ceil();
-
-    if (instruction.dose_and_rate?.dose_range) {
-      const lowDose = instruction.dose_and_rate.dose_range.low.value || "0";
-      const highDose = instruction.dose_and_rate.dose_range.high.value || "0";
-      const avgDose = divide(add(lowDose, highDose), 2);
-      return round(multiply(avgDose, numberOfDoses));
-    }
-
-    return round(multiply(doseValue, numberOfDoses));
-  }
 
   // Mutation to create invoice automatically after dispensing
   const { mutate: createInvoice, isPending: isCreatingInvoice } = useMutation({
