+++ src/types/emr/medicationRequest/medicationRequest.ts
@@ -6,6 +6,8 @@ import {
 import { InventoryRead } from "@/types/inventory/product/inventory";
 import { ProductKnowledgeBase } from "@/types/inventory/productKnowledge/productKnowledge";
 import { UserReadMinimal } from "@/types/user/user";
+import { add, divide, isZero, multiply, round } from "@/Utils/decimal";
+import Decimal from "decimal.js";
 
 export const MEDICATION_REQUEST_STATUS_COLORS = {
   active: "primary",
@@ -841,4 +843,83 @@ export function displayMedicationName(
       : "") ||
     ""
   );
+}
+
+export function computeInitialQuantity(
+  medication: MedicationRequestRead,
+): string {
+  const instruction = medication.dosage_instruction[0];
+  if (!instruction) {
+    return "0";
+  }
+
+  const doseValue = instruction.dose_and_rate?.dose_quantity?.value;
+  if (!doseValue) {
+    return "0";
+  }
+
+  const unitCode = instruction.dose_and_rate?.dose_quantity?.unit?.code;
+  const nonVolumetric = ["{tbl}", "{count}"];
+
+  if (unitCode && !nonVolumetric.includes(unitCode)) {
+    return "";
+  }
+
+  if (instruction.as_needed_boolean) {
+    return round(doseValue);
+  }
+
+  const repeat = instruction.timing?.repeat;
+  if (!repeat?.bounds_duration || !repeat.period_unit) {
+    return doseValue;
+  }
+
+  const convertToHours = (value: string, unit: string) => {
+    switch (unit) {
+      case "h":
+        return new Decimal(value);
+      case "d":
+        return multiply(value, 24);
+      case "wk":
+        return multiply(value, 24 * 7);
+      case "mo":
+        return multiply(value, 24 * 30);
+      case "a":
+        return multiply(value, 24 * 365);
+      default:
+        return 0;
+    }
+  };
+
+  const { frequency = 1, period = "1", period_unit, bounds_duration } = repeat;
+
+  const totalDurationInHours = convertToHours(
+    bounds_duration.value,
+    bounds_duration.unit,
+  );
+  const periodInHours = convertToHours(period, period_unit);
+
+  if (periodInHours === 0) {
+    return doseValue;
+  }
+
+  const doseIntervalInHours = divide(periodInHours, frequency);
+
+  if (isZero(doseIntervalInHours)) {
+    return doseValue;
+  }
+
+  const numberOfDoses = divide(
+    totalDurationInHours,
+    doseIntervalInHours,
+  ).ceil();
+
+  if (instruction.dose_and_rate?.dose_range) {
+    const lowDose = instruction.dose_and_rate.dose_range.low.value || "0";
+    const highDose = instruction.dose_and_rate.dose_range.high.value || "0";
+    const avgDose = divide(add(lowDose, highDose), 2);
+    return round(multiply(avgDose, numberOfDoses));
+  }
+
+  return round(multiply(doseValue, numberOfDoses));
 }
